+++ src/app/components/profesionales/profesional.html
@@ -1,15 +1,21 @@
 <plex-layout [aspect]=8>
   <plex-layout-main>
-    <header justify>
-      <plex-title titulo="Registro de profesional"></plex-title>
-      <div>
-        <plex-badge type="success mr-2">
-          Fecha y hora: {{ lblTurno }}.
-        </plex-badge>
-        <plex-button type="danger" label="volver" (click)="volverProfesional()"></plex-button>
-      </div>
-    </header>
     <form #formNuevoProf='ngForm'>
+      <ng-container *ngIf="editable">
+        <plex-title justify titulo="Editar profesional" size="lg">
+          <plex-button type="success" label="Confirmar Datos" (click)="actualizar($event)" class="pull-right"
+                       validateForm="true"></plex-button>
+          <plex-button type="danger" label="volver" (click)="volverProfesional()"></plex-button>
+        </plex-title>
+      </ng-container>
+      <ng-container *ngIf="!editable && confirmar == false && !ocultarBtn">
+        <plex-title justify titulo="Registro de profesional" size="lg">
+          <plex-badge type="success mr-2">
+            Fecha y hora: {{ lblTurno }}.
+          </plex-badge>
+          <plex-button type="danger" label="volver" (click)="volverProfesional()"></plex-button>
+        </plex-title>
+      </ng-container>
       <fieldset class="mt-3">
         <plex-title class="mt-3" titulo="Datos generales" size="sm"></plex-title>
         <!-- contacto -->
@@ -156,8 +162,6 @@
                           [(ngModel)]="profesional.domicilios[1].codigoPostal"
                           [required]="profesional.domicilios[1].tipo !== 'profesional'"></plex-int>
               </div>
-
-
             </div>
           </div>
         </div>
@@ -214,8 +218,6 @@
                               <plex-datetime label="Fecha del Título" type="date" [(ngModel)]="profesional.formacionGrado[0].fechaTitulo" name='fechaTitulo'
                                   [required]="true"></plex-datetime>
                           </div> -->
-
-
           <div class="col-md-4">
             <div *ngIf='!confirmar'>
               <plex-select *ngIf="!showOtraEntidadFormadora" label="Entidad formadora"
